requirements.txt
Django==5.2.7
pandas==2.3.3
openpyxl==3.1.5
beautifulsoup4==4.14.2
lxml==6.0.2
python-magic==0.4.27
chardet==5.2.0
pytest==8.4.2
pytest-django==4.11.1
gunicorn==23.0.0
whitenoise==6.11.0
psycopg2-binary==2.9.11
celery==5.4.0
redis==5.0.8
django-celery-results==2.5.1
dj-database-url==2.2.0